refactor(rent-requests): use suspense when loading requests
