+++ src/app/main/landlord/rent-requests/components/RentRequestFetcher.tsx
@@ -1,0 +1,11 @@
+import { RentRequestService } from "@/services/RentRequestService";
+import { RENT_REQUEST_STATUS } from "@/types/rentRequestTypes";
+import RentRequestList from "./RentRequestList";
+
+export default async function RentRequestFetcher() {
+  const rentRequests = await RentRequestService.fetchRentRequestsByLandlord(
+    RENT_REQUEST_STATUS.PENDING
+  );
+
+  return <RentRequestList rentRequests={rentRequests.data!} />;
+}
